# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Agronom Bot Internal API
  version: "1.10.0"
  description: |
    Telegram-бот «Карманный агроном» — API v1.10.0.
    Fixes: +signature, +UNAUTHORIZED, +monthly limits, +X-API-Ver usage, +multipart required, clarified dosage units, +X-User-ID header, +ask_expert endpoint, +protocol fields category/status/waiting_days, +assistant chat/confirm_plan.
  contact:
    name: Ivan Gromov
    url: https://github.com/agronom-bot
    email: i.gromov@agrostore.ru
servers:
  - url: https://api.agronom.local
    description: Production
  - url: https://staging.agronom.local
    description: Staging
security:
  - ApiKeyAuth: []
    RequestSignAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    RequestSignAuth:
      type: apiKey
      in: header
      name: X-Req-Sign
    HmacAuth:
      type: apiKey
      in: header
      name: X-Sign
  parameters:
    ApiVersionHeader:
      name: X-API-Ver
      in: header
      required: true
      schema:
        type: string
        enum: [v1]
    UserIdHeader:
      name: X-User-ID
      in: header
      required: true
      schema:
        type: integer
    RequestTimestampHeader:
      name: X-Req-Ts
      in: header
      required: true
      schema:
        type: integer
        description: Unix timestamp (seconds)
    RequestNonceHeader:
      name: X-Req-Nonce
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    DsrDeleteRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
    DsrDeleteResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: deleted
    ConsentStatus:
      type: object
      required: [doc_type, doc_version, status, updated_at, source]
      properties:
        doc_type:
          type: string
          example: privacy
        doc_version:
          type: string
          example: "1.0"
        status:
          type: boolean
        updated_at:
          type: string
          format: date-time
        source:
          type: string
          example: api
    ConsentRequest:
      type: object
      required: [doc_type, doc_version]
      properties:
        doc_type:
          type: string
          example: privacy
        doc_version:
          type: string
          example: "1.0"
        source:
          type: string
          nullable: true
          example: api
        meta:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [NO_LEAF, LIMIT_EXCEEDED, GPT_TIMEOUT, BAD_REQUEST, UNAUTHORIZED, UPGRADE_REQUIRED, TOO_MANY_REQUESTS, SERVICE_UNAVAILABLE, FORBIDDEN]
        message:
          type: string
...
paths:
...
  /v1/users/{user_id}/export:
    get:
      summary: Export user data (ZIP)
      operationId: exportUserData
      description: "Экспортирует данные пользователя в ZIP-архив (data.json)."
      tags: [users]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
          HmacAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ZIP archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/{user_id}/consents:
    get:
      summary: List user consents
      operationId: listUserConsents
      description: "Возвращает текущий статус согласий пользователя."
      tags: [users]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Consent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsentStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/{user_id}/consents/accept:
    post:
      summary: Accept consent
      operationId: acceptUserConsent
      description: "Фиксирует согласие пользователя."
      tags: [users]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
      responses:
        '200':
          description: Consent accepted
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/{user_id}/consents/revoke:
    post:
      summary: Revoke consent
      operationId: revokeUserConsent
      description: "Отзывает согласие пользователя."
      tags: [users]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
      responses:
        '200':
          description: Consent revoked
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/dsr/delete_user:
    post:
      summary: Delete user data (DSR)
      operationId: deleteUserDsr
      description: "Каскадное удаление данных пользователя по DSR-запросу."
      tags: [users]
      security:
        - ApiKeyAuth: []
          RequestSignAuth: []
          HmacAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/RequestTimestampHeader'
        - $ref: '#/components/parameters/RequestNonceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DsrDeleteRequest'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DsrDeleteResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...